Further developed style sheet 2
Added a little animation to the pictures of the proffesor that causes them to scale
up when the mouse hovers over them. Also created a tool bar with the tools for the class
Made links maroon and when hovering over them they become gold. Made most of the text readable